Initial DB architecture and document updates
The initial DB architecture includes the static information
that will be used for the static views of the game. The events
processing is still a draft to be defined in the future.
Additional change adds background to the tech tree image, since
arrows could not be seen in the dark gh bg.